Javascript fix for codemirror and ckeditor. Forces a wait until the script are loaded by using event handlers.
forgot to remove alert I was using as a quick test

Make the ckeditor settings more dymanic. If ckeditor_config_js or ckeditor_location are not set it will place them where they would normally be after the base path you have set.

some clean up, matching the name jquery suggests.

redid the javascript so that it first only load ckeditor/codemirror .js once. It does this by assuming the location provide in the first instances is the location that the other instances of the same type will use. It still though loads unquie options, theming, and code type for the following ckeditors/codemirror.

cleaned things by making it so it just goes and finds the elments once, and doesn't try to find ckeditor.js/codemirror.js again if it gets called a second time after the first load up.