Deduce Devise scope from _current user. Fixes #1386, partly.
If you configure Devise with sign_out_all_scopes = false,
and authenticate with a scope that's different from the default one (like :admin) the sign out link did not work. This fixes that.

Unfortunately i did not figure out how to add a helper spec for this.

TODO: Add a logout_path and logout_method option to RailsAdmin.config and fallback to this code.