Fix static asset serving from the RailsAdmin engine gem for production environments
Currently the ActionDispatch::Static middleware is added randomly in the
stack of middleware. This means that other middleware can interrupt, which we
have seen happening in production.

This patch also makes sure the middleware is only added when config.serve_static_assets is
set to true in your application in config/environments/production.rb