partial fix for issue 148: improve required/optional status for relations
There were a few bugs in the required/optional status reporting for
associations.  This change fixes some of them but maybe not all.
Model validations (e.g. validates_numericality_of) weren't being
considered at all so a field which was nullable in the db but which
had a validation that required the field would show up as "optional".
The check for belongs_to relations was backwards so belongs_to
relations which were nullable in the db would show up as required and
vice versa (again regardless of the model-level validations).

It turns out that the logic in base.rb was better than the logic in
the association subclasses that overrode it, so this is one of those
rare bugs that actually works better by deleting code.

This change adds a new table "rel_tests" which is used only by the
specs and doesn't make any sense in the context of the baseball app so
I added an initializer to dummy_app so it doesn't show up when you run
the app.