remove hard-coded widths on the .action and .other classes
These are the "EDIT/DELETE" actions and the "..." icon in the list
table.  Removing the width allows the browser to flow them closer
together if it needs to, so they take up less screen real estate and
leave more for the content.  They've still got 9px of padding so
they're plenty far apart from one another.