fix date format l10n bug in javascript date picker
Our form view code was using a date format of dd/MM/yyyy for the en_US
format but here in the US of A we put the month first, so if I picked
November 2 in the date picker it would get silently changed to
February 11 when the form was submitted.  I changed the format to an
iso8601-ish format so it looks awkward to everyone (except nerds) but
it should work everywhere.