add a "home" link at the top of each page that links to '/'
I noticed that when I added Rails Admin to my app that there was no
way to navigate back up and out.  This checkin adds a "home" link that
goes to '/' which in most cases should get people out of Rails admin
and back into the app that contains it.