issue #291: change name of error handling method from "render_error" to "handle_save_error"
"render_error" is a special name for rails, so calling our function by
that name caused it to interfere with Rails' error handling process.
Our function is a very special-case function designed to handle the
case where an object save fails for some reason.  It's not a
general-purpose error handling method, so in many cases if an error
happened our render_error would itself crash and hide the original
error.

It's better to let crashes percolate up to the application's error
handling page (or Rails' default), which this change enables.